Implemented behaviour to update player position and when a player joins/leaves

+++ Lastand-Core/src/utils.h
@@ -11,7 +11,7 @@ template <typename T>
 std::ostream &operator<<(std::ostream &os, const std::vector<T> &v) {
     os << "{ ";
     for (auto i: v)
-        os << i << ' ';
+        os << (int)i << ' ';
     os << '}';
     return os;
 }
